Rename to regex_iterator
